backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - using compatible versions for Python 3.13
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication and Security
pyjwt[crypto]==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.25.2

# Settings and Environment - using compatible versions
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2

aiosqlite==0.21.0
greenlet==3.2.4
